Major reordering - trying to move in a direction from coherence towards accuracy
This means no more Contemporary Rock (and subsequent removal of genres that aren't widely cited in most sources), and Metal is contained within Rock.

I'm not sure how to deal with Downtempo and Aesthetic yet - Downtempo still functions as an "Other" category, including miscellaneous Electronic stuff, such as Experimental Avant-Garde Early Pioneers stuff, Easy Listening and Chillout, Ambient, Chiptune or Video Game Music and finally some IDM-related music.

Aesthetic, on the other hand, is difficult to deal with for other reasons. It's so new that not many sources talk about it yet. I've noticed that Wikipedia now has a category for this, and calls it Hauntology, which is interesting. But like Downtempo, this is just Electronic stuff that isn't EDM. One solution I thought of was to rename Dance to Electronic. But then Industrial and Hip Hop would have to be included in there too, which I don't want. Besides, this would also complicate the fact that Industrial lies between Rock and Electronic, with the Gothic genres ending up in some weird in-between limbo.

The final thing I don't like a whole lot is the categorisation of Classical and Folk. On the biggest scale, all music is classified into two categories - either Art music or Vernacular music. Art music comprises Classical as well as historical musical traditions of other cultures. Vernacular music comprises Popular music (everything from Blues to Dance) as well as Folk music. Problem is, my Folk category also contains Art music from other cultures, making the entire genre problematic. And then Downtempo doesn't fit into it either. You could say that around half of the genres in Downtempo are part of Art music, as a third Experimental category. But then the rest are just super chill and loungy Electronic stuff that doesn't fit in Dance, but is most certainly a part of Popular music.

So where does this all leave us? I don't know, but classifiying music genres has never been an exact science. All of this is socially constructed, and that means there's a lot of muddiness and grey areas all over the place. The battle between accuracy and coherence is an unwinnable one. If the site was fully accurate, everything would be a total unnavigable mess. If the site was fully coherent, then the classification would be inaccurate and full of falsehoods.