Major simplification of the CSS - especially removing separate mobile div classes and using media queries instead.
I've also split the mobile genre grid into 2 columns to sort of match the look that Spotify has on mobile. This means I've had to descrease a lot of the sizing, so I'm not enitrely sure how it will look on a real mobile device yet. Will check after comitting this.